backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
yfinance
pandas==2.2.3
numpy==1.26.4
requests==2.31.0
lxml==5.2.1
beautifulsoup4==4.12.3
gunicorn==21.2.0
google-generativeai 
pennylane 
pennylane-lightning